requirements.txt
asgiref==3.11.1
audioread==3.1.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
decorator==5.2.1
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
dj-database-url==2.2.0
gunicorn==21.2.0
idna==3.12
imageio-ffmpeg==0.5.1
joblib==1.5.3
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.47.0
msgpack==1.1.2
musicbrainzngs==0.7.1
numba==0.65.0
numpy==2.4.4
packaging==26.1
pillow==12.2.0
platformdirs==4.9.6
pooch==1.9.0
psycopg2-binary==2.9.12
pyacoustid==1.3.1
pycparser==3.0
PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
soundfile==0.13.1
soxr==1.0.0
spotipy==2.26.0
sqlparse==0.5.5
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
whitenoise==6.7.0